web/requirements.txt
kafka-python==2.0.2
flask==3.0.0
Flask-WTF==1.2.1
flask_socketio==5.3.6